@@ -1,25 +1,26 @@
 {
   "name": "managed-timeout",
-  "version": "2.1.0",
+  "version": "3.0.0",
   "description": "setTimeout class wrapper with abilities to pause/resume, finish preemptively, and inspection",
   "repository": {
     "type": "git",
     "url": "https://github.com/religiosa1/managed-timeout.git"
   },
-  "main": "./dist/timeout.js",
-  "es2015": "./dist/timeout.mjs",
-  "types": "./dist/timeout.d.ts",
+  "main": "./dist/Timeout.js",
+  "es2015": "./dist/Timeout.mjs",
+  "types": "./dist/Timeout.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/timeout.mjs",
-      "require": "./dist/timeout.js",
-      "types": "./dist/timeout.d.ts"
+      "import": "./dist/Timeout.mjs",
+      "require": "./dist/Timeout.js",
+      "types": "./dist/Timeout.d.ts"
     }
   },
   "scripts": {
     "build": "tsup --dts",
     "test": "jest --coverage",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "prepack": "npm run lint && npm run build && npm run test"
   },
   "keywords": [
     "timer",
@@ -34,13 +35,14 @@
   ],
   "license": "ISC",
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
-    "eslint": "^8.2.0",
-    "eslint-plugin-jest": "^25.2.4",
-    "jest": "^27.3.1",
-    "mockdate": "^3.0.5",
-    "tsup": "^6.5.0",
-    "typescript": "^4.5.2"
+    "@types/jest": "^29.5.5",
+    "@typescript-eslint/eslint-plugin": "^6.8.0",
+    "@typescript-eslint/parser": "^6.8.0",
+    "eslint": "^8.51.0",
+    "eslint-plugin-jest": "^27.4.2",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "tsup": "^7.2.0",
+    "typescript": "^5.2.2"
   }
 }
