@@ -1,10 +1,9 @@
-const { Timeout } = require("../dist/timeout");
-const { Timer } = require("./Timer");
+import { Timeout } from "../src/Timeout";
 const delay = 1000;
 
 describe("repeated calls", ()=>{
   it("returns false on cancelled timeout", () => {
-    let to = new Timeout(()=>{}, delay);
+    const to = new Timeout(()=>{}, delay);
     to.cancel();
     expect(to.pause()).toBe(false);
     expect(to.resume()).toBe(false);
@@ -14,9 +13,8 @@ describe("repeated calls", ()=>{
   });
 
   it("returns false on ended timeout", () => {
-    let timer = new Timer();
-    let to = new Timeout(()=>{}, delay);
-    timer.advance(delay);
+    const to = new Timeout(()=>{}, delay);
+    jest.advanceTimersByTime(delay);
     expect(to.pause()).toBe(false);
     expect(to.resume()).toBe(false);
     expect(to.cancel()).toBe(false);
@@ -25,19 +23,18 @@ describe("repeated calls", ()=>{
   });
 
   test("resume does nothing on nonpaused timeout", () => {
-    let to = new Timeout(()=>{}, delay);
-    let st = to.resume();
+    const to = new Timeout(()=>{}, delay);
+    const st = to.resume();
     expect(st).toBe(false);
     expect(to.paused).toBe(false);
     expect(to.isPending).toBe(true);
   });
 
   test("pause does nothing on paused timeout", () => {
-    let timer = new Timer();
-    let to = new Timeout(()=>{}, delay);
+    const to = new Timeout(()=>{}, delay);
     to.pause();
-    timer.advance(delay * 0.5);
-    let st = to.pause();
+    jest.advanceTimersByTime(delay * 0.5);
+    const st = to.pause();
     expect(st).toBe(false);
     expect(to.paused).toBe(true);
     expect(to.isPending).toBe(true);
